dino_chrome_bot/requirements.txt
selenium
opencv-python
numpy
pyautogui
pillow
webdriver-manager